refactored fetch fasade
