+++ backend/src/shared/README.md
@@ -1,0 +1,3 @@
+# Readme
+
+Files here are shared between the frontend and backend like "result" and the types for payloads and responses
