+++ backend/src/shared/api/types/PlayerSit.ts
@@ -1,0 +1,9 @@
+export type PlayerSitPayload = {
+  selectedSeatNumber: string;
+  socketId: string;
+};
+
+export type PlayerSitResult = {
+  seatNumber: string;
+  socketId: string;
+};
