+++ frontend/src/components/Table/Seat/Seat.tsx
@@ -1,6 +1,9 @@
 import React, { useCallback } from 'react';
 import { Socket } from 'socket.io-client';
+
 import FetchFasade from '../../../fetch/FetchFasade';
+
+import { PlayerSitPayload, PlayerSitResult } from '../../../../../backend/src/shared/api/types/PlayerSit';
 
 type SeatProps = {
   seatNumber: string;
@@ -11,7 +14,8 @@ type SeatProps = {
 export default function Seat({ seatNumber, chipCount, socket }: SeatProps) {
   const playerSit = useCallback(async (event: React.MouseEvent) => {
     const payload = { selectedSeatNumber: event.currentTarget.id, socketId: socket.id };
-    const result = await FetchFasade.post('/api/actions/playerSit', payload);
+
+    const result = await FetchFasade.post<PlayerSitPayload, PlayerSitResult>('/api/actions/playerSit', payload);
 
     if (result.ok) {
       console.log(result.getValue());
